:art: Improve structure / format of the code：カスタムフックとtypeファイルを作成し依存関係の整理
