+++ types/type.ts
@@ -1,0 +1,6 @@
+export type Gpa = {
+  subject: string;
+  A: number;
+  B?: number;
+  fullMark: number;
+};
